[fix] model module별로 file  분리
