+++ README.md
@@ -28,3 +28,7 @@ insertMoney
 4. insertMoney 횟수만큼 기록을 한다 [O]
 
 순서데로 테스트 코드 진행 
+
+### 뷰를 나눌 필요성 
+나눠져서 각각해당되는 부분만 관리 하는 것이 더 깔끔해보인다. Why not 모델도 나눠져 있어서 더 흐름 보기 용이 
+-> vendingMachineView / WalletView
